+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'quave:collections',
-  version: '2.0.0',
+  version: '2.0.1',
   summary: 'Utility package to create Meteor collections in a standard way',
   git: 'https://github.com/quavedev/collections',
 });
@@ -12,18 +12,17 @@ Npm.depends({
 });
 
 Package.onUse(function (api) {
-  api.versionsFrom('METEOR@3.0-alpha.18');
+  api.versionsFrom('3.0-beta.7');
 
   api.use([
-    'ecmascript@0.16.7||0.16.8-alpha300.18',
-    'mongo@1.0.0||2.0.0||2.0.0-alpha300.17',
-    'minimongo@1.9.3||2.0.0-alpha300.18',
-    'ejson@1.1.3||1.1.4-alpha300.18',
+    'ecmascript@0.16.7||0.16.8-beta300.7',
+    'mongo@1.0.0||2.0.0||2.0.0-beta300.7',
+    'minimongo@1.9.3||2.0.0-beta300.7',
+    'ejson@1.1.3||1.1.4-beta300.7',
   ]);
   api.imply('mongo');
 
   api.use('raix:eventemitter@1.0.0');
-  api.use('tmeasday:check-npm-versions@1.0.2');
 
   api.use('quave:settings@1.0.0');
 
@@ -31,7 +30,7 @@ Package.onUse(function (api) {
 });
 
 Package.onTest(function (api) {
-  api.use(['ecmascript', 'tinytest', 'insecure', 'autopublish', 'mongo']);
+  api.use(['ecmascript@0.16.7||0.16.8-beta300.7', 'tinytest', 'insecure', 'autopublish', 'mongo@1.0.0||2.0.0||2.0.0-beta300.7']);
 
   api.addFiles(['helpers.js', 'helpers-test.js']);
 });